cc_get_tweet/requirements.txt
future
requests
requests_oauthlib
python-twitter
pytz

